==> Code/Graph Convolutional Network Pipeline/1. Doc2vec.ipynb <==
import pandas as pd
import numpy as np

from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from nltk.tokenize import word_tokenize

from sklearn.metrics.pairwise import cosine_similarity

df_info = pd.read_csv('node_information.csv', header=None, names=['ID', 'Year', 'Title', 'Author', 'Journal', 'Abstract'])

# Doc2Vec

# Build a corpus
data = df_info['Title'] + ' ' + df_info['Abstract']

# Tokenization
tagged_data = [TaggedDocument(words=word_tokenize(_d.lower()), tags=[str(i)]) for i, _d in enumerate(data)]

# Hyper-parameters
max_epochs = 50
vec_size = 50
alpha = 0.01

# Build a model
model = Doc2Vec(vector_size=vec_size,
                window=3,
                alpha=alpha, 
                min_alpha=0.0025,
                min_count=1,
                workers=16)

model.build_vocab(tagged_data)

# Training
for epoch in range(max_epochs):
    print('iteration {0}'.format(epoch))
    model.train(tagged_data,
                total_examples=model.corpus_count,
                epochs=model.epochs)
    # decrease the learning rate
    model.alpha -= 0.0002
    # fix the learning rate, no decay
    model.min_alpha = model.alpha

model.save("d2v.model")
print("Model Saved")

# Most similar documents
similar_doc = model.docvecs.most_similar('1')
print(similar_doc)

# Get document vectors
model.docvecs[1]

# Create a vector dataframe
df_vector = df_info.copy()
m_v = np.array([model.docvecs[i] for i in range(len(df_info))])
m_v = pd.DataFrame(m_v)
m_v.columns = ['v_' + str(x) for x in m_v.columns]

df_vector = pd.concat([df_vector[['ID']], m_v], axis=1)

# Save the vector
df_vector.to_csv('doc2vec.csv', index=None)

# Cosine Similarity

df_vector = pd.read_csv('doc2vec.csv')

# Generate a similarity matrix
sim = cosine_similarity(df_vector.iloc[:,1:])
df_sim = pd.DataFrame(sim, index=df_vector.iloc[:,0], columns=df_vector.iloc[:,0])
