==> Code/Graph Convolutional Network Pipeline/2. Feature Engineering.ipynb <==
import networkx as nx
import pandas as pd

# 1. Data Preperation

# Read all the files
df_train = pd.read_csv('training_set.txt', header=None, sep=' ', names=['X1', 'X2', 'Y'])
df_test = pd.read_csv('testing_set.txt', header=None, sep=' ', names=['X1', 'X2'])
df_info = pd.read_csv('node_information.csv', header=None, names=['ID', 'Year', 'Title', 'Author', 'Journal', 'Abstract'])

df_aut = pd.read_csv('df_info_author_centrality.csv')
df_doc2vec = pd.read_csv('doc2vec.csv')

# Build a graph for feature extraction
edges = df_train[df_train['Y']==1]
G = nx.from_pandas_edgelist(edges, 'X1', 'X2', True, create_using=nx.DiGraph())

# 2. Node Features

## 2.1 Doc2Vec

# Get embeddings for titles and abstracts
df_info = df_info.merge(df_doc2vec, left_on='ID', right_on='ID', how='left')

# 3. Save to the disk

## 3.1 Delete Redundencies

cols_del = ['Title', 'Author', 'Journal', 'Abstract']
df_info = df_info.drop(cols_del, axis=1)

## 3.2 Save

# Save them to the disk
df_info.to_csv('df_features.csv', index=None)
